pyproject.toml
[tool.poetry]
name = "kg-rag"
version = "0.1.0"
description = ""
authors = ["eros28095821 <eros28095821@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
neo4j = "^5.27.0"
sentence-transformers = "^3.3.1"
torch = "^2.5.1"
pydantic = ">2.7"
flask = "^3.1.0"
langchain-ollama = "^0.2.1"
numpy = "1.26.4"
prompt-toolkit = "^3.0.48"
prompt = "^0.4.1"
gradio = "^5.8.0"
fpdf2 = "^2.8.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
